refactor(scripts): promote timeMedian/median/round1 into shared bench-timing lib
Forge phase 28 extracted timeMedian in token-benchmark.ts, already wired
within that file. The codebase-wide duplicate scan found median()/round1()
duplicated verbatim across query-benchmark.ts, incremental-benchmark.ts
(twice), and benchmark.ts, plus the "loop N times, time each run, median()"
pattern repeated 8 more times across those files.

Promotes median/round1/timeMedian into scripts/lib/bench-timing.ts
(matching the existing bench-config.ts/fork-engine.ts shared-module
pattern) and wires median/round1 into all 4 scripts. Adopts timeMedian at
the 3 call sites that were already unconditionally async with no new
async boundary (incremental-benchmark.ts fullBuildMs/noopRebuildMs worker,
benchmark.ts noopRebuildMs worker).

Left 5 remaining duplicate sites unconverted since adopting timeMedian
there requires making currently-synchronous functions async, which is a
control-flow change rather than a mechanical adoption -- tracked in #1774.

Impact: 4 functions changed, 14 affected
